primary example improvements: more opts, more sequences, latest clippy (#69)

+++ q_compress_cli/src/arrow_number_like.rs
@@ -82,6 +82,6 @@ impl ArrowNumberLike for TimestampNanos {
   }
 
   fn to_arrow(self) -> <Self::ArrowPrimitive as ArrowPrimitiveType>::Native {
-    self.to_total_parts() as i64
+    self.to_total_parts()
   }
 }
